Fixed Powershell and winget scripts for building and packaging to work on Windows
* Inno Setup's package name for winget was incorrect.
* Inno Setup 6 does not use the standard install locations. It uses
AppData/Local/Programs. Added that to the Prefixes in
Install-BuildDependencies script.
* Also Get-Command fails because the way iscc location is added to Path
is for session. Inno Setup installer doesn't add iscc to PATH, so it's
never found and Packaging with -BuildInstaller never works. In this case
, added registering the winget's package $Binary, if it doesn't already
exist in PATH, since we most likely want to have it accessible from
terminal/powershell.
* Moved Cmake Install