Various
* Fixed a bug on the early exit check in Update
* Added a helper to update the text of the QLineEdit without emitting
signals, which caused a corruption of data in the config_data object.
* Now checking if config value and QLineEdit text value are equal before
trying to overwrite any of the config values. No unnecessary overwrites.
* Fixed bug where we used setPlainText to assign the QLineEdit and after
that the setPlaceholderText which didn't set the value of the
QLineEdits. Now using the correct function, including blocking the
singals when first setting the widgets' values in SetSettings.